requirements.txt
zarr
hdbscan
umap-learn
networkx
tqdm
torch 
esm
faiss-cpu
pandas
scikit-learn